feat: add support for array-based command setting when running containers via docker run
Previously, all commands had to be quoted, as they were passed as is to an entrypoint that only understood a single argument- /bin/bash -l -c. This caused issues with generic usage of docker run where the command often isn't quoted.

This change allows folks to use either quoted commands (existing) or an array of commands (new). It does so by treating the passed arguments as a single string in all cases via $*.

See https://stackoverflow.com/a/64689634 and https://unix.stackexchange.com/a/129077 for more details on how variable expansion works in shell.

Closes #658