feat(test): gracefully cleanup failed docker build
Some docker `simple_build` (or `build_with_build_time_env_vars`) create
additional resources as part of their process. This is for the dbs, they
create a volume and a network. But if the build itself fails, the
created resources never reach the cleanup logic. This commit defers the
unwrapping of the build results, to the caller; allowing us to inject
cleanup logic where needed