feat: add debug option to Docker build commands
This would enable an interative shell to be opened up if a build fails. This is still experimental
and I'm running into some issues using this with orbstack, so this PR is still very much a WIP.

https://github.com/orbstack/orbstack/issues/1700

Without this option, the only way to debug a failing build is to `--out` the build, insert multi-stage
build commands in the dockerfile and then use `--target` to get in-progress container ID to play around
with, which is super lame. We need a better way.